New: mei.js allows specification of arguments on subscription
The user can now specify arguments when subscribing, not just when publishing [i.e. mei.Events.subscribe(topic, callback, arguments)]. This is to allow subscribing a box to perform a specific action when an event is triggered.

Prior to this functionality, if the user wished to have multiple boxes execute tasks, it was necessary to publish multiple events with one set of arguments per box. Now, we can make multiple subscriptions so that each box can perform separate tasks when an event is triggered.

(Subscribing 'createBoxWorker' to an event with no arguments does not specify much about what should happen when the event is triggered, hence this functionality).